performanceTest/requirements.txt
altair==4.1.0
pae==1.0.0
requests==2.24.0
tensorflow==2.2.0
numpy==1.19.0
pandas==1.0.5
Pillow==7.2.0
altair_saver==0.5.0